Today
I know it’s not a lot, but I’ve been searching the internet as well as
trying many different things to get this to work. For some reason in my
accept_f_name() it won’t update the tableWidget. It gets to the
accept_f_name(), but won’t touch the table. Maybe you can figure it out