DESCRIPTION
This should make the file. It will not make the damn file. I know I am
getting the methods right. I couted it and they’re there. It WILL NOT
create the file on my computer. I made the file to see if it would
write to it. It wouldn’t.  I tried stream, QFileDialog, and
QFile+QTestStream and nothing is working. Can you try it on your
computer?